+++ derive/src/utils/mod.rs
@@ -10,6 +10,7 @@ pub mod interface_attr;
 pub mod interface_hash;
 pub mod macros;
 pub mod meta_match;
+pub mod remove_attr;
 pub mod rename_rule;
 pub mod type_utils;
 pub mod with_arg;